feat: add tampermonkey config

+++ scripts/extension-config-backup.sh
@@ -3,6 +3,8 @@
 # alias ebk="sh ~/Codespaces/my-macos-build/scripts/extension-config-backup.sh"
 FILE_UBLOCK=$(find ~/Downloads -maxdepth 1 -name 'my-ublock-backup*.txt' | head -n1)
 FILE_UBLOCK_NAME=ublock-advanced.txt
+FILE_TAMPERMONKEY=$(find ~/Downloads -maxdepth 1 -name 'tampermonkey-backup-*.zip' | head -n1)
+FILE_TAMPERMONKEY_NAME=tampermonkey-backup.zip
 FILE_VIOLENTMONKEY=$(find ~/Downloads -maxdepth 1 -name 'violentmonkey_*.zip' | head -n1)
 FILE_VIOLENTMONKEY_NAME=violentmonkey-backup.zip
 FILE_TONGWENTANG=$(find ~/Downloads -maxdepth 1 -name 'tongwentang-pref*.json' | head -n1)
@@ -22,7 +24,21 @@ if [ -f "$FILE_UBLOCK" ]; then
   echo "done."
   echo ""
 fi
-# find scripts_*.zip and backup
+# find tampermonkey-backup-**.zip and backup
+if [ -f "$FILE_TAMPERMONKEY" ]; then
+  mv "$FILE_TAMPERMONKEY" "$PROJECT_DIR$FILE_TAMPERMONKEY_NAME"
+  cd $PROJECT_DIR || exit
+  git add $FILE_TAMPERMONKEY_NAME
+  echo ""
+  printf 'Find TAMPERMONKEY Configuration: %s\n' "$FILE_TAMPERMONKEY"
+  printf '     --> tampermonkey-backup.zip\n\n'
+  printf 'Backup...'
+  git commit -q -m "feat: Update tampermonkey Configuration by ebk"
+  git push -q
+  echo "done."
+  echo ""
+fi
+# find violentmonkey_*.zip and backup
 if [ -f "$FILE_VIOLENTMONKEY" ]; then
   mv "$FILE_VIOLENTMONKEY" "$PROJECT_DIR$FILE_VIOLENTMONKEY_NAME"
   cd $PROJECT_DIR || exit
